Bodge together EXT-X-PROGRAM-DATE-TIME support
Activated if the H264 video stream contains pic_timing data in the SEI
headers.  This also has the effect of causing timestamps to be measured
in 90ths of a second since the UTC epoch, rather than simply being
derived from the (unwrapped) PTS/DTS values from the source stream.

Does not handle leap-seconds, because of course not.